fixed PR review
