+++ backend/ctfd/plugin/admin/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/admin/__init__.py
