+++ backend/ctfd/plugin/admin/controllers/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/admin/controllers/__init__.py
