+++ backend/ctfd/plugin/admin/routes/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/admin/routes/__init__.py
