--- backend/ctfd/plugin/controllers/__init__.py
@@ -1,1 +1,0 @@
-# /plugin/controllers/__init__.py
