--- backend/ctfd/plugin/models/__init__.py
@@ -1,9 +1,0 @@
-# /plugin/models/__init__.py
-
-from .World import World
-from .User import User
-from .TeamMember import TeamMember
-from .Team import Team
-from .enums import TeamRole
-
-__all__ = ["World", "Team", "User", "TeamMember", "TeamRole"]
