+++ backend/ctfd/plugin/team/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/team/__init__.py
