+++ backend/ctfd/plugin/team/controllers/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/team/controllers/__init__.py
