+++ backend/ctfd/plugin/team/models/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/team/models/__init__.py
