+++ backend/ctfd/plugin/team/models/enums.py
@@ -1,5 +1,5 @@
 """
-# /plugin/models/enums.py
+# /plugin/team/models/enums.py
 Enumerations to ensure type safety and prevent magic string bugs
 """
