+++ backend/ctfd/plugin/team/routes/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/team/routes/__init__.py
