+++ backend/ctfd/plugin/user/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/user/__init__.py
