+++ backend/ctfd/plugin/user/controllers/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/user/controllers/__init__.py
