+++ backend/ctfd/plugin/user/models/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/user/models/__init__.py
