+++ backend/ctfd/plugin/user/routes/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/user/routes/__init__.py
