+++ backend/ctfd/plugin/world/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/world/__init__.py
