+++ backend/ctfd/plugin/world/controllers/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/world/controllers/__init__.py
