+++ backend/ctfd/plugin/world/models/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/world/models/World.py
