+++ backend/ctfd/plugin/world/routes/__init__.py
@@ -1,0 +1,1 @@
+# /plugin/world/routes/__init__.py
